requirements.txt
discord.py
python-dotenv
feedparser
apscheduler
flask
yt-dlp